+++ order-service/src/main/java/shop/genieus/order/adapter/in/api/dto/response/PaymentResponse.java
@@ -1,0 +1,9 @@
+package shop.genieus.order.adapter.in.api.dto.response;
+
+import shop.genieus.order.domain.model.entity.Order;
+
+public record PaymentResponse(Long orderId) {
+  public static PaymentResponse toResponse(Order order) {
+    return new PaymentResponse(order.getOrderId());
+  }
+}
