+++ order-service/src/main/java/shop/genieus/order/adapter/in/event/external/OrderKafkaListener.java
@@ -1,0 +1,23 @@
+package shop.genieus.order.adapter.in.event.external;
+
+import com.genieus.common.event.util.EventRouter;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.kafka.support.KafkaHeaders;
+import org.springframework.messaging.handler.annotation.Header;
+import org.springframework.messaging.handler.annotation.Payload;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class OrderKafkaListener {
+
+  private final EventRouter eventRouter;
+
+  @KafkaListener(topics = "${spring.kafka.consumer.topic.payment}")
+  public void consume(@Header(KafkaHeaders.RECEIVED_TOPIC) String topic, @Payload String payload) {
+    eventRouter.route(topic, payload);
+  }
+}
