+++ order-service/src/main/java/shop/genieus/order/adapter/in/scheduler/OrderScheduler.java
@@ -1,0 +1,19 @@
+package shop.genieus.order.adapter.in.scheduler;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
+import shop.genieus.order.application.in.scheduler.OrderSchedulerService;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class OrderScheduler {
+  private final OrderSchedulerService schedulerService;
+
+  @Scheduled(fixedDelayString = "${order.scheduler.poll-interval:30000}")
+  public void pollExpiredMessages() {
+    schedulerService.popExpiredOrders();
+  }
+}
