+++ order-service/src/main/java/shop/genieus/order/global/config/FeignConfig.java
@@ -4,5 +4,5 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@EnableFeignClients(basePackages = "shop.genieus.order.infrastructure.client.*")
+@EnableFeignClients(basePackages = "shop.genieus.order.adapter.out.client.*")
 public class FeignConfig {}
