--- order-service/src/main/java/shop/genieus/order/infrastructure/client/coupon/CouponFeignClient.java
@@ -1,7 +1,0 @@
-package shop.genieus.order.infrastructure.client.coupon;
-
-import com.genieus.common.internal.client.CouponInternalClient;
-import org.springframework.cloud.openfeign.FeignClient;
-
-@FeignClient(name = "coupon-service")
-public interface CouponFeignClient extends CouponInternalClient {}
