--- order-service/src/main/java/shop/genieus/order/infrastructure/client/payment/PaymentFeignClient.java
@@ -1,7 +1,0 @@
-package shop.genieus.order.infrastructure.client.payment;
-
-import com.genieus.common.internal.client.PaymentInternalClient;
-import org.springframework.cloud.openfeign.FeignClient;
-
-@FeignClient(name = "payment-service")
-public interface PaymentFeignClient extends PaymentInternalClient {}
