--- order-service/src/main/java/shop/genieus/order/infrastructure/event/external/producer/OrderKafkaProducer.java
@@ -1,54 +1,0 @@
-package shop.genieus.order.infrastructure.event.external.producer;
-
-import com.genieus.common.event.DomainEvent;
-import com.genieus.common.event.EventEnvelope;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.support.SendResult;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class OrderKafkaProducer {
-
-  private final KafkaTemplate<String, EventEnvelope<? extends DomainEvent>> kafkaTemplate;
-
-  @Value("${spring.kafka.template.default-topic}")
-  private String orderTopic;
-
-  public void publish(String key, EventEnvelope<? extends DomainEvent> eventEnvelop) {
-    ProducerRecord<String, EventEnvelope<? extends DomainEvent>> record =
-        new ProducerRecord<>(orderTopic, key, eventEnvelop);
-    kafkaTemplate
-        .send(record)
-        .whenComplete((sendResult, ex) -> handleSendResult(eventEnvelop, sendResult, ex));
-  }
-
-  public void publish(EventEnvelope<? extends DomainEvent> eventEnvelop) {
-    ProducerRecord<String, EventEnvelope<? extends DomainEvent>> record =
-        new ProducerRecord<>(orderTopic, eventEnvelop);
-    kafkaTemplate
-        .send(record)
-        .whenComplete((sendResult, ex) -> handleSendResult(eventEnvelop, sendResult, ex));
-  }
-
-  private void handleSendResult(
-      EventEnvelope<? extends DomainEvent> eventEnvelope,
-      SendResult<String, EventEnvelope<? extends DomainEvent>> sendResult,
-      Throwable ex) {
-    if (ex != null) {
-      log.error("Kafka 이벤트 전송 실패: eventEnvelope={}", eventEnvelope, ex);
-      return;
-    }
-    log.info(
-        "Kafka 이벤트 전송 성공: key={}, value={}",
-        sendResult.getProducerRecord().key() != null
-            ? sendResult.getProducerRecord().key()
-            : "null",
-        sendResult.getProducerRecord().value());
-  }
-}
