--- order-service/src/main/java/shop/genieus/order/infrastructure/persistence/command/OrderCommandAdapter.java
@@ -1,37 +1,0 @@
-package shop.genieus.order.infrastructure.persistence.command;
-
-import static shop.genieus.order.global.exception.CustomNotFoundException.*;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import shop.genieus.order.application.out.persistence.OrderCommandPort;
-import shop.genieus.order.domain.model.entity.Order;
-import shop.genieus.order.infrastructure.persistence.command.repository.OrderJpaRepository;
-
-@Component
-@RequiredArgsConstructor
-public class OrderCommandAdapter implements OrderCommandPort {
-  private final OrderJpaRepository orderJpaRepository;
-
-  @Override
-  public Order save(Order order) {
-    return orderJpaRepository.save(order);
-  }
-
-  @Override
-  public Order findById(Long orderId) {
-    return orderJpaRepository.findById(orderId).orElseThrow(OrderNotFoundException::new);
-  }
-
-  @Override
-  public List<Order> findAll(List<Long> orderIds) {
-    return orderJpaRepository.findAllById(orderIds);
-  }
-
-  @Override
-  public void expireAll(List<Long> orderIds, LocalDateTime expiredAt) {
-    orderJpaRepository.expireAll(orderIds, expiredAt);
-  }
-}
